Reorganized shiva integration tests a bit.
