+++ package.json
@@ -23,11 +23,9 @@
     "shiva-test": "npm run shiva-eslint && ASTRAL_ROOT=\"$(pwd)/lib/\" NODE_ENV=test lab -m 1000 -v -c -a code test/shiva/unit test/shiva/functional",
     "shiva-unit": "ASTRAL_ROOT=\"$(pwd)/lib/\" NODE_ENV=test lab -v -c -a code test/shiva/unit",
     "shiva-functional": "ASTRAL_ROOT=\"$(pwd)/lib/\" NODE_ENV=test lab -v -a code test/shiva/functional",
-    "shiva-integration": "ASTRAL_ROOT=\"$(pwd)/lib/\" NODE_ENV=test lab -m 900000 -v -a code integration/shiva",
-    "shiva-integration-aws": "ASTRAL_ROOT=\"$(pwd)/lib/\" NODE_ENV=test lab -a code -v -m 300000 integration/shiva/aws.js",
+    "shiva-integration": "ASTRAL_ROOT=\"$(pwd)/lib/\" LOG_LEVEL=trace NODE_ENV=test lab -a code -v -m 300000 integration/shiva/tasks integration/shiva/models",
     "shiva-integration-models": "ASTRAL_ROOT=\"$(pwd)/lib/\" NODE_ENV=test lab -a code -v -m 300000 integration/shiva/models",
-    "shiva-integration-server": "ASTRAL_ROOT=\"$(pwd)/lib/\" LOG_LEVEL=trace NODE_ENV=test lab -a code -v -m 300000 integration/shiva/server.js | bunyan",
-    "shiva-integration-asg": "ASTRAL_ROOT=\"$(pwd)/lib/\" LOG_LEVEL=trace NODE_ENV=test lab -a code -v -m 300000 integration/shiva/auto-scaling-groups.js | bunyan",
+    "shiva-integration-tasks": "ASTRAL_ROOT=\"$(pwd)/lib/\" LOG_LEVEL=trace NODE_ENV=test lab -a code -v -m 300000 integration/shiva/tasks | bunyan",
     "metis-start": "node lib/metis/index.js",
     "metis-eslint": "eslint lib/metis/",
     "metis-test": "npm run metis-eslint && ASTRAL_ROOT=\"$(pwd)/lib/\" NODE_ENV=test lab -m 1000 -v -c -a code test/metis",
